+++ frontend/src/@hooks/kkogkkog/useChangeKkogKkogStatus.ts
